+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.x branch
 ## 2.6 branch
 ### 2.6.4
+* added tests for lower dependencies;
 * improved exception message when no binary file is found;
 * no longer uses the `Folder` class.